requirements.txt
streamlit
pillow
agno
duckduckgo-search
google-genai
ddgs
python-dotenv
